+++ javelin/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Release 0.1.3.0
 
 * Improved performance for the `aggregateWith` function.
+* Documentation improvements.
 
 ## Release 0.1.2.0